chore: move useClickOutside into hook folder
